backend/requirements.txt
fastapi[standard]
uvicorn
google-genai
google-cloud-vision
elevenlabs
httpx
pydantic-settings
python-dotenv
